backend/requirements.txt
django>=5.2
pillow>=10.0.0
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.5.0
selenium>=4.30.0
django-cors-headers>=4.3.1
django-widget-tweaks>=1.5.0
django-redis>=5.4.0
coverage>=7.4.1
pytz>=2024.1
python-dotenv>=1.0.0
# Testing
pytest>=8.0.0
pytest-django>=4.8.0
pytest-cov>=4.1.0
# Development
black>=24.3.0
flake8>=7.0.0
isort>=5.13.2
